validação de rotas privadas

+++ src/app/useCases/articles/deleteArticle.ts
@@ -1,9 +1,16 @@
 import { Request, Response } from "express";
 import { Article } from "../../models/Article";
+import { isValidObjectId } from "mongoose";
 
 export async function deleteArticle(req: Request, res: Response) {
   try {
     const { articleId } = req.params;
+
+    const isValidId = isValidObjectId(articleId);
+
+    if (!isValidId) {
+      return res.status(400).json({ error: "O Id informado não é válido!" });
+    }
 
     await Article.findByIdAndDelete(articleId);
